First pass of multi-colour

+++ pkg/chip8/instruction_test.go
@@ -414,9 +414,9 @@ func vmForTest() *VM {
 	for y := 0; y < DisplayHeight; y++ {
 		for x := 0; x < DisplayWidth; x++ {
 			r := rand.Intn(100)
-			pixel := uint8(0)
+			pixel := uint16(0)
 			if r > 50 {
-				pixel = uint8(1)
+				pixel = uint16(1)
 			}
 			v.display[x][y] = pixel
 		}
